add form validation

+++ app/routes/admin/brand/$slug.delete.tsx
@@ -2,21 +2,45 @@ import { Brand } from "@prisma/client";
 import { useParams } from "react-router-dom";
 import {
     ActionFunction,
-    Form,
     LoaderFunction,
     redirect,
+    useActionData,
     useCatch,
     useLoaderData,
 } from "remix";
 
 import prisma from "~/db";
 import { Sidebar } from "~/modules/brand";
-import { Text, Button } from "~/modules/ui";
+import { Form, Button } from "~/modules/ui";
+import { GetActionData } from "~/types";
 
-export const action: ActionFunction = async ({ request }) => {
+function validateUuid(name: unknown) {
+    if (typeof name !== "string" || name.length < 1) {
+        return [`Form not submitted correctly`];
+    }
+
+    return [];
+}
+
+type ActionData = GetActionData<"uuid">;
+
+export const action: ActionFunction = async ({
+    request,
+}): Promise<Response | ActionData> => {
     const body = new URLSearchParams(await request.text());
 
     const uuid = body.get("uuid") ?? "";
+
+    const fields = { uuid };
+
+    const formError = validateUuid(uuid);
+
+    if (formError) {
+        return {
+            formError,
+            fields,
+        };
+    }
 
     await prisma.brand.delete({
         where: {
@@ -34,7 +58,7 @@ type LoaderData = {
 export const loader: LoaderFunction = async ({ params }) => {
     const brand = await prisma.brand.findFirst({
         where: {
-            name: params.slug,
+            slug: params.slug,
         },
     });
 
@@ -48,11 +72,13 @@ export const loader: LoaderFunction = async ({ params }) => {
 };
 
 export default function Edit() {
+    const actionData = useActionData<ActionData>();
     const { brand } = useLoaderData<LoaderData>();
+
     return (
         <Sidebar title={` Delete ${brand.name}`}>
             <p>Are you sure that you want to delete {brand.name}?</p>
-            <Form replace method="post">
+            <Form replace method="post" errors={actionData?.formError}>
                 <input type="hidden" name="uuid" value={brand.uuid} />
                 <div className="max-w-md pt-8">
                     <Button type="submit">Delete {brand.name}</Button>
@@ -72,7 +98,6 @@ export function CatchBoundary() {
             return (
                 <Sidebar title={`The brand ${params.slug} does not exist`} />
             );
-
         default:
             throw new Error(
                 `Unexpected caught response with status: ${caught.status}`
